+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1025](https://github.com/spegel-org/spegel/pull/1025) Add unit tests for serving manifests through mirror.
 - [#1026](https://github.com/spegel-org/spegel/pull/1026) Switch to using a generic functional options package.
 - [#1031](https://github.com/spegel-org/spegel/pull/1031) Report errors on debug web view.
+- [#1034](https://github.com/spegel-org/spegel/pull/1034) Refactor to use shared reference in image and distribution path.
   
 ### Deprecated